% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca.R
\name{compute_pca}
\alias{compute_pca}
\title{Compute a PCA via svd of the cov.}
\usage{
compute_pca(data, features = NULL)
}
\arguments{
\item{data}{a data frame containing numeric features}

\item{features}{vector of feature columns (must be numeric)}
}
\value{
a pca object to use for (re-)transformation
}
\description{
Compute a PCA via svd of the cov.
}
\examples{
   test_data <- as.data.frame(cbind(1:4, 4:7))
   colnames(test_data) <- c("feat1", "feat2")
   test_pca <- compute_pca(test_data)

}
