% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigentransform.R
\name{eigentransform}
\alias{eigentransform}
\title{Apply a PCA to data.}
\usage{
eigentransform(data, pca)
}
\arguments{
\item{data}{a data frame with data to be transformed.}

\item{pca}{the pca object used for transformation.}
}
\value{
transformed data (data.frame)
}
\description{
Apply a PCA to data.
}
\examples{
\dontrun{
  data_trafo <- eigentransform( data, pca )
}

}
