% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca.R
\name{plot.pca}
\alias{plot.pca}
\title{Quick bivariate plot of PCA data.}
\usage{
\method{plot}{pca}(pca, cex = NULL, col = NULL, pch = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{pca}{a pca object (see above)}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Just trying some S3 method overloading,
with defaults for some crucial plot attributes.
see here:
https://stackoverflow.com/q/13120895
https://aksela.wordpress.com/2018/08/18/simple-example-for-creating-a-custom-s3-class-with-methods
}
