% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrafo.R
\name{retrafo}
\alias{retrafo}
\title{Reverse transform loadings to data with the PCA.}
\usage{
retrafo(data_trafo, pca)
}
\arguments{
\item{data_trafo}{data.frame points in pca space}

\item{pca}{the pca object defining transformed space.}
}
\value{
retransformed data (data.frame)
}
\description{
Reverse transform loadings to data with the PCA.
}
\examples{
\dontrun{
  data_retrafo <- retrafo( data_trafo, pca )
}

}
