% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Apply a PCA to data.}
\usage{
transform(data, pca)
}
\arguments{
\item{data}{a data frame with data to be transformed.}

\item{pca}{the pca object used for transformation.}
}
\value{
transformed data (data.frame)
}
\description{
Apply a PCA to data.
}
\examples{
\dontrun{
  data_trafo <- transform( data, pca )
}

}
