% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca.R
\name{plot.pca}
\alias{plot.pca}
\title{Quick bivariate plot of PCA data.}
\usage{
\method{plot}{pca}(
  x,
  y = NULL,
  cex = NULL,
  col = NULL,
  pch = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{a pca object (see above).}

\item{y}{not used; needed for S3 signature.}

\item{cex}{marker size}

\item{col}{marker fill color}

\item{pch}{shape of the scatter marker}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{...}{this is an ellipsis, passed to \code{plot.default}.}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Just trying some S3 method overloading,
with defaults for some crucial plot attributes.
see here:
https://stackoverflow.com/q/13120895
https://aksela.wordpress.com/2018/08/18/simple-example-for-creating-a-custom-s3-class-with-methods
https://www.datamentor.io/r-programming/s3-class
}
